refactor cohort builder filter panels
