+++ packages/frontend/src/features/CohortBuilder/ExplorerTable/ExplorerTable.tsx
@@ -165,9 +165,9 @@ const ExplorerTable = ({ index, tableConfig }: ExplorerTableProps) => {
     const totalRowCount = tableConfig?.pageLimit
       ? Math.min(
           pageLimit,
-          data?.data._aggregation?.[index]._totalCount ?? pagination.pageSize,
+          data?.data?._aggregation?.[index]._totalCount ?? pagination.pageSize,
         )
-      : (data?.data._aggregation?.[index]._totalCount ?? pagination.pageSize);
+      : (data?.data?._aggregation?.[index]._totalCount ?? pagination.pageSize);
     const limitLabel = tableConfig?.pageLimit
       ? (tableConfig?.pageLimit?.label ?? DEFAULT_PAGE_LIMIT_LABEL)
       : 'Rows per Page:';
