+++ packages/frontend/src/features/CohortBuilder/Panels/types.ts
@@ -1,0 +1,10 @@
+import { TabsConfig } from '../types';
+import { FacetDefinition, FacetType } from '@gen3/core';
+import { FacetDataHooks } from '../../../components/facets';
+
+export interface TabbablePanelProps {
+  filters: TabsConfig;
+  tabTitle: string;
+  facetDefinitions: Record<string, FacetDefinition>;
+  facetDataHooks: Record<FacetType, FacetDataHooks>;
+}
